requirements.txt
pandas
numpy
langchain
langchain-community
langchain-google-genai
chromadb
sentence-transformers
pypdf
python-dotenv